+++ ChessGame/Pawn.cs
@@ -52,7 +52,7 @@ namespace ChessGame
             p = new DataStructure.point(possition.i + 1, possition.j - 1);
             if (Validation.IsValidPosition(p))
             {
-                if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                 {
                     possiblePossitions.Add(p);
                 }
@@ -64,7 +64,7 @@ namespace ChessGame
             p = new DataStructure.point(possition.i + 1, possition.j + 1);
             if (Validation.IsValidPosition(p))
             {
-                if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                 {
                     possiblePossitions.Add(p);
                 }
@@ -110,7 +110,7 @@ namespace ChessGame
             p = new DataStructure.point(possition.i - 1, possition.j - 1);
             if (Validation.IsValidPosition(p))
             {
-                if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                 {
                     possiblePossitions.Add(p);
                 }
@@ -122,7 +122,7 @@ namespace ChessGame
             p = new DataStructure.point(possition.i - 1, possition.j + 1);
             if (Validation.IsValidPosition(p))
             {
-                if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                 {
                     possiblePossitions.Add(p);
                 }
