+++ ChessGame/Queen.cs
@@ -31,7 +31,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -47,7 +47,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -63,7 +63,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -79,7 +79,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -100,7 +100,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -120,7 +120,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -140,7 +140,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -160,7 +160,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -191,7 +191,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -207,7 +207,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -223,7 +223,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -239,7 +239,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -260,7 +260,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -280,7 +280,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -300,7 +300,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
@@ -320,7 +320,7 @@ namespace ChessGame
                     }
                     else
                     {
-                        if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                        if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                         {
                             possiblePossitions.Add(p);
                         }
