+++ ChessGame/Rock.cs
@@ -31,7 +31,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -47,7 +47,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -63,7 +63,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -79,7 +79,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.white, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.white, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -103,7 +103,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -119,7 +119,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -135,7 +135,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
@@ -151,7 +151,7 @@ namespace ChessGame
                 }
                 else
                 {
-                    if (Validation.IsOpponent(p, DataStructure.color.black, boardMatrix))
+                    if (Validation.IsOpponent(p, (int)DataStructure.color.black, boardMatrix))
                     {
                         possiblePossitions.Add(p);
                     }
